requirements.txt
flask==3.0.3
flask-cors==4.0.1
cryptography==42.0.8
python-dotenv==1.0.1
